make GrantableEntity inherit from Entity instead of a mixin, create the _safe_grant method with hopefully sensible checking logic, add more unit tests
